Get a table of fragments based on a peptide sequence.

fragments() Generates a table of fragments from a given peptide character string.

Usage

fragments(
  sequence = "SAMPLE",
  type = c("y", "b"),
  charge = c(1, 2, 3, 4),
  loss = NULL
)

Arguments

sequence

The character string representing a peptide, or poly amino acid. The canonical 20 amino acids are encoded in and chemical modifications can be represented by and floating point numerical value enclosed by square brackets. If a canonical amino acid is also enclosed in the square brackets [M15.99] it is assumed that the numerical value is in addition to the mass of the residue, and thus represents a post-translational modification (PTM). Any value, in a square bracket is assumed to represent a monomer in the chain [57.021], and reflected in the fragment ion series.

Example: SAMPLER, SA[M15.99]PLER

type

The ion fragmentation type to calculate the mass values for.

Allowable: y, b

Default: c("y", "b")

charge

The specified fragment ion charges to calculate.

Allowable: 1:4

Default: c(1, 2, 3, 4)

loss

The specified neutral mass loss experienced in some mass spectrometers.

Allowable: water, amine

Default: none

Value

The output is a tibble of predicted mass values based on the input parameters. The table consists of the columns

   `ion`: the ion type (y,b) and the charge represented by +

   `mz`: the calculated m/z value of the ion

   `z`: the charge state of the ion

   `seq`: the partial sequence represented by the fragmentation

   `pair`: a tag use to identify which ions originated from the y/b ion split

   `pos`: the position at which the fragmentation occurred

   `type`: the ion type

Examples

fragments("SAMPLER", charge = 1)
#>     ion        mz z     seq pair pos      type
#> 1   b1+  88.03931 1       S  p01   1         b
#> 2   b2+ 159.07643 1      SA  p02   2         b
#> 3   y1+ 175.11894 1       R  p06   1         y
#> 4   b3+ 290.11691 1     SAM  p03   3         b
#> 5   y2+ 304.16153 1      ER  p05   2         y
#> 6   b4+ 387.16967 1    SAMP  p04   4         b
#> 7  MH++ 402.20763 2 SAMPLER  p00  NA precursor
#> 8   y3+ 417.24559 1     LER  p04   3         y
#> 9   b5+ 500.25373 1   SAMPL  p05   5         b
#> 10  y4+ 514.29836 1    PLER  p03   4         y
#> 11  b6+ 629.29632 1  SAMPLE  p06   6         b
#> 12  y5+ 645.33884 1   MPLER  p02   5         y
#> 13  y6+ 716.37596 1  AMPLER  p01   6         y
#> 14  MH+ 803.40799 1 SAMPLER  p00  NA precursor